\name{providedHook1}
\alias{providedHook1}
\title{Provide a chunk hook for knitr to be executed by run_hooks.}
\usage{
providedHook1(before, options, envir)
}
\arguments{
  \item{before}{}

  \item{options}{}

  \item{envir}{}
}
\description{
Provide a chunk hook for knitr to be executed by run_hooks.
}

